fix test issus
