+++ panel_chat_examples/__init__.py
@@ -1,22 +1,7 @@
-"""Example recipes for Panel ChatInterface."""
+"""Shared functionality for Panel Chat Examples."""
+from panel_chat_examples._environment_widget import EnvironmentWidgetBase
 
 __version__ = "0.0.0"
 
 
-def add(a: int, b: int) -> int:
-    """
-    Add two integers.
-
-    Args:
-        a:
-            The first operand.
-        b:
-            The second operand.
-
-    Examples:
-        Add two integers
-
-            r = add(2, 3)
-            print(r)  # 5
-    """
-    return a + b
+__all__ = ["EnvironmentWidgetBase"]
